removed exec-htap-params, incorporated exec-htap.sh changes

+++ quickstart/exec-htap.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
-
-source config.sh
-
+BIN_DIR="$( dirname "${BASH_SOURCE[0]}" )"
+pushd "$BIN_DIR" > /dev/null
+# Default values
+IM=0	# Throughput Monitoring Interval in milliseconds
+SW=300	# Sampling window
+source ./config.sh
+# Allow overriding parameters on the command line as KEY=VALUE
+for param in "$@"
+do
+   eval $param
+done
 echo "execute htap-$SCALE benchmark, time: $TIME  tpcc: $CWORKERS  chbenchmark: $HWORKERS"
-
 # Write config file (use Mac/Linux compatible 'sed -i')
 cp template-config.xml config.xml
 sed -i.bak -e "s%###SPLICE_URL###%jdbc:splice://$SPLICE_HOST:$SPLICE_PORT/splicedb%g" config.xml
@@ -12,18 +19,15 @@ sed -i.bak -e "s%###TIME###%$TIME%g" config.xml
 sed -i.bak -e "s%###WORKERS_TPCC###%$CWORKERS%g" config.xml
 sed -i.bak -e "s%###WORKERS_TPCH###%$HWORKERS%g" config.xml
 rm -rf config.xml.bak
-
-pushd ..
-
-# always rebuild - changing DDL or dialect files requires rebuild of oltpbenchmark
-rm -rf build
-ant
-
+cd ..
+# Changing DDL or dialect files requires rebuild of oltpbenchmark
+if [[ "$PREBUILT" != "true" ]]; then
+  rm -rf build
+  ant
+fi
 mkdir -p results
-
 session=htap-${SCALE}_${CWORKERS}_${HWORKERS}
-./oltpbenchmark -b 'tpcc,chbenchmark' -c quickstart/config.xml --execute=true -s 300 -ss -o $session | tee results/$session.out
-
+./oltpbenchmark -b 'tpcc,chbenchmark' -c quickstart/config.xml --execute=true -im $IM -s $SW -ss -o $session | tee results/$session.out
 popd
-
 echo 'done'
+
